requirements.txt
Babel==2.5.1
blinker==1.4
boto==2.34.0
chardet==2.3.0
Cheetah==2.4.4
click==6.7
cloud-init==0.7.6
configobj==5.0.6
dominate==2.3.1
ecdsa==0.11
euca2ools==3.1.0
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
jsonpatch==1.3
jsonpointer==1.0
lxml==3.4.0
Markdown==2.6.9
MarkupSafe==1.0
MySQL-python==1.2.3
oauth==1.0.1
paramiko==1.15.1
passlib==1.6.1
prettytable==0.7.2
pycrypto==2.6.1
pyserial==2.6
python-apt==0.9.3.12
python-ldap==2.4.10
pytz==2017.3
PyYAML==3.11
requestbuilder==0.2.3
requests==2.4.3
six==1.8.0
SQLAlchemy==1.1.15
urllib3==1.9.1
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
